refactor(useResize): 리사이즈 로직 useResize hook으로 분리
